pyproject.toml
[project]
name = "summarizeandmailraindroplinks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.2",
    "openai>=1.57.0",
    "pytest>=8.4.2",
    "python-dateutil>=2.9.0.post0",
    "readability-lxml>=0.8.1",
    "sendgrid>=6.11.0",
]
